refactor(#278): seperate useMouseTrackingPreview hooks
