+++ apps/frontend/src/features/memo/components/MemoPreview.tsx
@@ -1,49 +1,19 @@
-import { useRef, useEffect } from 'react';
+import { useEffect, useRef } from 'react';
 import { useStore } from '@xyflow/react';
+import { useMouseTrackingPreview } from '@/features/drawing/hooks/useMouseTrackingPreview';
 
 export const MemoPreview = () => {
-  const divRef = useRef<HTMLDivElement>(null);
   const zoom = useStore((s) => s.transform[2]);
   const zoomRef = useRef(zoom);
-  const frameRef = useRef<number | null>(null);
-  const latestPositionRef = useRef<{ x: number; y: number } | null>(null);
 
   useEffect(() => {
     zoomRef.current = zoom;
   }, [zoom]);
 
-  useEffect(() => {
-    const el = divRef.current;
-    if (!el) return;
-
-    const handleMouseMove = (e: MouseEvent) => {
-      latestPositionRef.current = { x: e.clientX, y: e.clientY };
-
-      if (frameRef.current !== null) return;
-
-      frameRef.current = window.requestAnimationFrame(() => {
-        frameRef.current = null;
-        const position = latestPositionRef.current;
-        if (!position) return;
-
-        el.style.display = 'block';
-        el.style.transform = `translate3d(${position.x}px, ${
-          position.y - 48
-        }px, 0) scale(${zoomRef.current})`;
-      });
-    };
-
-    window.addEventListener('mousemove', handleMouseMove);
-
-    return () => {
-      window.removeEventListener('mousemove', handleMouseMove);
-
-      if (frameRef.current !== null) {
-        window.cancelAnimationFrame(frameRef.current);
-        frameRef.current = null;
-      }
-    };
-  }, []);
+  const divRef = useMouseTrackingPreview(
+    (position) =>
+      `translate3d(${position.x}px, ${position.y - 48}px, 0) scale(${zoomRef.current})`,
+  );
 
   return (
     <div
